migrate test_haproxyconn to pytest
